Refactor allocation front files
* Extract components from AllocationPage.jsx and AllocationSummaryTable.jsx
* New AllocationList.jsx component
* New AllocationObservationPlansTable.jsx component
* New AllocationSummary.jsx component
* New SimpleMenu.jsx component
* Break down the components in allocation/ into two different directories `AllocationPage` and `AllocationSummary`
* New util.jsx file to avoid duplication
* Ignore `import/prefer-default-export` in eslint
