+++ static/js/components/allocation/util.jsx
@@ -1,0 +1,54 @@
+import CircularProgress from "@mui/material/CircularProgress";
+import React from "react";
+import { createTheme } from "@mui/material/styles";
+
+export function getAllocationTitle(allocation, instrumentList, telescopeList) {
+  const { instrument_id } = allocation;
+  const instrument = instrumentList?.filter((i) => i.id === instrument_id)[0];
+
+  const telescope_id = instrument?.telescope_id;
+  const telescope = telescopeList?.filter((t) => t.id === telescope_id)[0];
+
+  if (!(instrument?.name && telescope?.name)) {
+    return (
+      <div>
+        <CircularProgress color="secondary" />
+      </div>
+    );
+  }
+
+  return `${instrument?.name}/${telescope?.nickname}`;
+}
+
+// Tweak responsive styling
+export const getTweakedMuiTheme = (theme) =>
+  createTheme({
+    palette: theme.palette,
+    components: {
+      MUIDataTablePagination: {
+        styleOverrides: {
+          toolbar: {
+            flexFlow: "row wrap",
+            justifyContent: "flex-end",
+            padding: "0.5rem 1rem 0",
+            [theme.breakpoints.up("sm")]: {
+              // Cancel out small screen styling and replace
+              padding: "0px",
+              paddingRight: "2px",
+              flexFlow: "row nowrap"
+            }
+          },
+          tableCellContainer: {
+            padding: "1rem"
+          },
+          selectRoot: {
+            marginRight: "0.5rem",
+            [theme.breakpoints.up("sm")]: {
+              marginLeft: "0",
+              marginRight: "2rem"
+            }
+          }
+        }
+      }
+    }
+  });
